+++ integrations/supabase/src/haystack_integrations/components/retrievers/supabase/embedding_retriever.py
@@ -31,7 +31,9 @@ class SupabasePgvectorEmbeddingRetriever(PgvectorEmbeddingRetriever):
     ```python
     from haystack import Document, Pipeline
     from haystack.document_stores.types.policy import DuplicatePolicy
-    from haystack.components.embedders import SentenceTransformersTextEmbedder, SentenceTransformersDocumentEmbedder
+    # Requires: pip install sentence-transformers-haystack
+    from haystack_integrations.components.embedders.sentence_transformers import SentenceTransformersTextEmbedder
+    from haystack_integrations.components.embedders.sentence_transformers import SentenceTransformersDocumentEmbedder
 
     from haystack_integrations.document_stores.supabase import SupabasePgvectorDocumentStore
     from haystack_integrations.components.retrievers.supabase import SupabasePgvectorEmbeddingRetriever
@@ -47,7 +49,6 @@ class SupabasePgvectorEmbeddingRetriever(PgvectorEmbeddingRetriever):
                  Document(content="In certain places, you can witness the phenomenon of bioluminescent waves.")]
 
     document_embedder = SentenceTransformersDocumentEmbedder()
-    document_embedder.warm_up()
     documents_with_embeddings = document_embedder.run(documents)
     document_store.write_documents(documents_with_embeddings.get("documents"), policy=DuplicatePolicy.OVERWRITE)